強制ターン終了をゲーム進行に組み込んだ
